update userSlice to manage signUp API call
