+++ src/utilities/redux/slices/users/userSlice.tsx
@@ -1,6 +1,8 @@
-import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { UiSliceInterface, UserInterface } from '../../../types/UserInterface';
+import { createSlice, createAsyncThunk, PayloadAction } from '@reduxjs/toolkit';
+import { SignUpInterface, UiSliceInterface, UserInterface } from '../../../types/UserInterface';
+import { signUp } from '../../../api/users';
 import { RootState } from '../../store';
+
 
 const initialState: UiSliceInterface = {
   auth: {
@@ -19,8 +21,25 @@ const initialState: UiSliceInterface = {
   },
   status: {
     isAuthenticated: false,
+    isLoading: false,
+    isSuccess: false,
+    isError: {
+      status: false,
+      message: ''
+    }
   },
 };
+
+export const register = createAsyncThunk(
+  'sign-up',
+  async (user: SignUpInterface, thunkAPI) => {
+    try {
+      return await signUp(user)
+    } catch (error) {
+      return thunkAPI.rejectWithValue('Unable to register!')
+    }
+  }
+)
 
 const usersSlice = createSlice({
   name: 'ui',
@@ -29,9 +48,31 @@ const usersSlice = createSlice({
     setAuthUser: (state, action: PayloadAction<UserInterface>) => {
       state.auth.user = action.payload;
     },
+    reset: (state) => {
+      state.status.isLoading = false;
+      state.status.isSuccess = false;
+      state.status.isError = { status: false };
+    }
+    
   },
+  extraReducers: (builder) => {
+    builder
+      // REGISTER
+      .addCase(register.pending, (state) => {
+        state.status.isLoading = true;
+      })
+      .addCase(register.fulfilled, (state, action) => {
+        state.status.isLoading = false;
+        state.status.isSuccess = true;
+        state.auth.user = action.payload;
+      })
+      .addCase(register.rejected, (state) => {
+        state.status.isLoading = false;
+        state.status.isError = { status: true };
+      })
+  }
 });
 
 export const selectAuthUser = (state: RootState) => state.ui.auth.user;
-export const { setAuthUser } = usersSlice.actions;
+export const { setAuthUser, reset } = usersSlice.actions;
 export default usersSlice.reducer;
